added tabular regression and renamed structured data clf widget
