+++ js/src/lib/components/Icons/IconTabularClassification.svelte
