+++ js/src/lib/components/Icons/IconTabularRegression.svelte
@@ -1,0 +1,21 @@
+<script lang="ts">
+	export let classNames = "";
+</script>
+
+<svg 
+    class={classNames}
+    xmlns="http://www.w3.org/2000/svg" 
+    xmlns:xlink="http://www.w3.org/1999/xlink" 
+    aria-hidden="true" 
+    role="img" 
+    class="iconify iconify--carbon" 
+    width="1em" 
+    height="1em" 
+    preserveAspectRatio="xMidYMid meet" 
+    viewBox="0 0 32 32">
+
+    <path 
+        fill="currentColor" 
+        d="m4.67 28l6.39-12l7.3 6.49a2 2 0 0 0 1.7.47a2 2 0 0 0 1.42-1.07L27 10.9l-1.82-.9l-5.49 11l-7.3-6.49a2 2 0 0 0-1.68-.51a2 2 0 0 0-1.42 1L4 25V2H2v26a2 2 0 0 0 2 2h26v-2Z"
+        />
+</svg>
